+++ Client/Controllers/UserController.cs
@@ -14,22 +14,38 @@ namespace UserClient.Controllers
     public class UserController : Controller
     {
         private IUserService _userService;
+        private readonly IHttpContextAccessor _contextAccessor;
 
-        public UserController(IUserService userService)
+        public UserController(IHttpContextAccessor contextAccessor, IUserService userService)
         {
             _userService = userService;
+            _contextAccessor = contextAccessor;
         }
 
         // GET: User
         [AuthorizeForScopes(ScopeKeySection = "User:UserScope")]
         public async Task<ActionResult> Index()
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             return View(await _userService.GetAsync());
         }
 
         // GET: User/Details/5
         public async Task<ActionResult> Details(string id)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             return View(await _userService.GetAsync(id));
         }
 
@@ -56,6 +72,13 @@ namespace UserClient.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind("Oid,FirstName,Surname,Username")] User user)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             await _userService.AddAsync(user);
             return RedirectToAction("Index");
         }
@@ -63,6 +86,13 @@ namespace UserClient.Controllers
         // GET: User/Edit/5
         public async Task<ActionResult> Edit(string id)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             User user = await this._userService.GetAsync(id);
 
             if (user == null)
@@ -79,6 +109,13 @@ namespace UserClient.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(string id, [Bind("Id,Oid,FirstName,Surname,Username,IsPaid,IsAdmin,IsDeleted")] User user)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             await _userService.EditAsync(id, user);
             return RedirectToAction("Index");
         }
@@ -86,6 +123,13 @@ namespace UserClient.Controllers
         // GET: User/Delete/5
         public async Task<ActionResult> Delete(string id)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             User user = await this._userService.GetAsync(id);
 
             if (user == null)
@@ -101,8 +145,22 @@ namespace UserClient.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Delete(string id, [Bind("Id,Oid,FirstName,Surname")] User user)
         {
+            var isAdmin = await IsAdministrator();
+
+            if (!isAdmin)
+            {
+                return NotFound();
+            }
+
             await _userService.DeleteAsync(id);
             return RedirectToAction("Index");
+        }
+
+        private async Task<bool> IsAdministrator()
+        {
+            var identity = _contextAccessor.HttpContext.User.Identity as System.Security.Claims.ClaimsIdentity;
+            var user = await _userService.GetAsync(identity.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier).Value);
+            return user.IsAdmin;
         }
     }
 }
